+++ NEWS.md
@@ -1,5 +1,21 @@
 # fluxfixer (development version)
 
+## Minor changes
+
+* Better final message when setting `skip_sapflow_calc` as `TRUE` in
+ `run_fluxfixer()`
+
+* Better message when setting `modify_z` as `TRUE` in
+ `run_fluxfixer()`
+
+
+## Bug fixes
+
+* QC flag determination in `run_fluxfixer()` now works correctly.
+
+* Z-score modification during short attenuation periods in `run_fluxfixer()`
+ now works correctly.
+
 # fluxfixer 1.0.0
 
 * First CRAN release